improve: move compare resource version methods to internal utils
This should not be user facing. At least not in any obvious scenerio.
