+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/ReconcilerUtilsInternal.java
@@ -31,6 +31,7 @@ import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import io.javaoperatorsdk.operator.api.reconciler.Constants;
 import io.javaoperatorsdk.operator.api.reconciler.ControllerConfiguration;
+import io.javaoperatorsdk.operator.api.reconciler.NonComparableResourceVersionException;
 import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 
 @SuppressWarnings("rawtypes")
@@ -240,5 +241,124 @@ public class ReconcilerUtilsInternal {
       }
     }
     return false;
+  }
+
+  /**
+   * Compares resource versions of two resources. This is a convenience method that extracts the
+   * resource versions from the metadata and delegates to {@link
+   * #validateAndCompareResourceVersions(String, String)}.
+   *
+   * @param h1 first resource
+   * @param h2 second resource
+   * @return negative if h1 is older, zero if equal, positive if h1 is newer
+   * @throws NonComparableResourceVersionException if either resource version is invalid
+   */
+  public static int validateAndCompareResourceVersions(HasMetadata h1, HasMetadata h2) {
+    return validateAndCompareResourceVersions(
+        h1.getMetadata().getResourceVersion(), h2.getMetadata().getResourceVersion());
+  }
+
+  /**
+   * Compares the resource versions of two Kubernetes resources.
+   *
+   * <p>This method extracts the resource versions from the metadata of both resources and delegates
+   * to {@link #compareResourceVersions(String, String)} for the actual comparison.
+   *
+   * @param h1 the first resource to compare
+   * @param h2 the second resource to compare
+   * @return a negative integer if h1's version is less than h2's version, zero if they are equal,
+   *     or a positive integer if h1's version is greater than h2's version
+   * @see #compareResourceVersions(String, String)
+   */
+  public static int compareResourceVersions(HasMetadata h1, HasMetadata h2) {
+    return compareResourceVersions(
+        h1.getMetadata().getResourceVersion(), h2.getMetadata().getResourceVersion());
+  }
+
+  /**
+   * Compares two resource version strings using a length-first, then lexicographic comparison
+   * algorithm.
+   *
+   * <p>The comparison is performed in two steps:
+   *
+   * <ol>
+   *   <li>First, compare the lengths of the version strings. A longer version string is considered
+   *       greater than a shorter one. This works correctly for numeric versions because larger
+   *       numbers have more digits (e.g., "100" > "99").
+   *   <li>If the lengths are equal, perform a character-by-character lexicographic comparison until
+   *       a difference is found.
+   * </ol>
+   *
+   * <p>This algorithm is more efficient than parsing the versions as numbers, especially for
+   * Kubernetes resource versions which are typically monotonically increasing numeric strings.
+   *
+   * <p><strong>Note:</strong> This method does not validate that the input strings are numeric. For
+   * validated numeric comparison, use {@link #validateAndCompareResourceVersions(String, String)}.
+   *
+   * @param v1 the first resource version string
+   * @param v2 the second resource version string
+   * @return a negative integer if v1 is less than v2, zero if they are equal, or a positive integer
+   *     if v1 is greater than v2
+   * @see #validateAndCompareResourceVersions(String, String)
+   */
+  public static int compareResourceVersions(String v1, String v2) {
+    int comparison = v1.length() - v2.length();
+    if (comparison != 0) {
+      return comparison;
+    }
+    for (int i = 0; i < v2.length(); i++) {
+      int comp = v1.charAt(i) - v2.charAt(i);
+      if (comp != 0) {
+        return comp;
+      }
+    }
+    return 0;
+  }
+
+  /**
+   * Compares two Kubernetes resource versions numerically. Kubernetes resource versions are
+   * expected to be numeric strings that increase monotonically. This method assumes both versions
+   * are valid numeric strings without leading zeros.
+   *
+   * @param v1 first resource version
+   * @param v2 second resource version
+   * @return negative if v1 is older, zero if equal, positive if v1 is newer
+   * @throws NonComparableResourceVersionException if either resource version is empty, has leading
+   *     zeros, or contains non-numeric characters
+   */
+  public static int validateAndCompareResourceVersions(String v1, String v2) {
+    int v1Length = validateResourceVersion(v1);
+    int v2Length = validateResourceVersion(v2);
+    int comparison = v1Length - v2Length;
+    if (comparison != 0) {
+      return comparison;
+    }
+    for (int i = 0; i < v2Length; i++) {
+      int comp = v1.charAt(i) - v2.charAt(i);
+      if (comp != 0) {
+        return comp;
+      }
+    }
+    return 0;
+  }
+
+  private static int validateResourceVersion(String v1) {
+    int v1Length = v1.length();
+    if (v1Length == 0) {
+      throw new NonComparableResourceVersionException("Resource version is empty");
+    }
+    for (int i = 0; i < v1Length; i++) {
+      char char1 = v1.charAt(i);
+      if (char1 == '0') {
+        if (i == 0) {
+          throw new NonComparableResourceVersionException(
+              "Resource version cannot begin with 0: " + v1);
+        }
+      } else if (char1 < '0' || char1 > '9') {
+        throw new NonComparableResourceVersionException(
+            "Non numeric characters in resource version: " + v1);
+      }
+    }
+    return v1Length;
   }
 }
