+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/ReconcilerUtilsInternalTest.java
@@ -17,7 +17,10 @@ package io.javaoperatorsdk.operator;
 
 import java.net.URI;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import io.fabric8.kubernetes.api.model.*;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
@@ -29,6 +32,7 @@ import io.fabric8.kubernetes.client.http.HttpRequest;
 import io.fabric8.kubernetes.model.annotation.Group;
 import io.fabric8.kubernetes.model.annotation.ShortNames;
 import io.fabric8.kubernetes.model.annotation.Version;
+import io.javaoperatorsdk.operator.api.reconciler.NonComparableResourceVersionException;
 import io.javaoperatorsdk.operator.sample.simple.TestCustomReconciler;
 import io.javaoperatorsdk.operator.sample.simple.TestCustomResource;
 
@@ -43,7 +47,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 class ReconcilerUtilsInternalTest {
-
+  private static final Logger log = LoggerFactory.getLogger(ReconcilerUtilsInternalTest.class);
   public static final String RESOURCE_URI =
       "https://kubernetes.docker.internal:6443/apis/tomcatoperator.io/v1/tomcats";
 
@@ -182,5 +186,136 @@ class ReconcilerUtilsInternalTest {
     public void setReplicas(Integer replicas) {
       this.replicas = replicas;
     }
+  }
+
+  // naive performance test that compares the work case scenario for the parsing and non-parsing
+  // variants
+  @Test
+  @Disabled
+  public void compareResourcePerformanceTest() {
+    var execNum = 30000000;
+    var startTime = System.currentTimeMillis();
+    for (int i = 0; i < execNum; i++) {
+      var res = ReconcilerUtilsInternal.compareResourceVersions("123456788" + i, "123456789" + i);
+    }
+    var dur1 = System.currentTimeMillis() - startTime;
+    log.info("Duration without parsing: {}", dur1);
+    startTime = System.currentTimeMillis();
+    for (int i = 0; i < execNum; i++) {
+      var res = Long.parseLong("123456788" + i) > Long.parseLong("123456789" + i);
+    }
+    var dur2 = System.currentTimeMillis() - startTime;
+    log.info("Duration with parsing:   {}", dur2);
+
+    assertThat(dur1).isLessThan(dur2);
+  }
+
+  @Test
+  void validateAndCompareResourceVersionsTest() {
+    assertThat(ReconcilerUtilsInternal.validateAndCompareResourceVersions("11", "22")).isNegative();
+    assertThat(ReconcilerUtilsInternal.validateAndCompareResourceVersions("22", "11")).isPositive();
+    assertThat(ReconcilerUtilsInternal.validateAndCompareResourceVersions("1", "1")).isZero();
+    assertThat(ReconcilerUtilsInternal.validateAndCompareResourceVersions("11", "11")).isZero();
+    assertThat(ReconcilerUtilsInternal.validateAndCompareResourceVersions("123", "2")).isPositive();
+    assertThat(ReconcilerUtilsInternal.validateAndCompareResourceVersions("3", "211")).isNegative();
+
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("aa", "22"));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("11", "ba"));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("", "22"));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("11", ""));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("01", "123"));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("123", "01"));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("3213", "123a"));
+    assertThrows(
+        NonComparableResourceVersionException.class,
+        () -> ReconcilerUtilsInternal.validateAndCompareResourceVersions("321", "123a"));
+  }
+
+  @Test
+  void compareResourceVersionsWithStrings() {
+    // Test equal versions
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("1", "1")).isZero();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("123", "123")).isZero();
+
+    // Test different lengths - shorter version is less than longer version
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("1", "12")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("12", "1")).isPositive();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("99", "100")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("100", "99")).isPositive();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("9", "100")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("100", "9")).isPositive();
+
+    // Test same length - lexicographic comparison
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("1", "2")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("2", "1")).isPositive();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("11", "12")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("12", "11")).isPositive();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("99", "100")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("100", "99")).isPositive();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("123", "124")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("124", "123")).isPositive();
+
+    // Test with non-numeric strings (algorithm should still work character-wise)
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("a", "b")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("b", "a")).isPositive();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("abc", "abd")).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("abd", "abc")).isPositive();
+
+    // Test edge cases with larger numbers
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("1234567890", "1234567891"))
+        .isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions("1234567891", "1234567890"))
+        .isPositive();
+  }
+
+  @Test
+  void compareResourceVersionsWithHasMetadata() {
+    // Test equal versions
+    HasMetadata resource1 = createResourceWithVersion("123");
+    HasMetadata resource2 = createResourceWithVersion("123");
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource1, resource2)).isZero();
+
+    // Test different lengths
+    resource1 = createResourceWithVersion("1");
+    resource2 = createResourceWithVersion("12");
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource1, resource2)).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource2, resource1)).isPositive();
+
+    // Test same length, different values
+    resource1 = createResourceWithVersion("100");
+    resource2 = createResourceWithVersion("200");
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource1, resource2)).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource2, resource1)).isPositive();
+
+    // Test realistic Kubernetes resource versions
+    resource1 = createResourceWithVersion("12345");
+    resource2 = createResourceWithVersion("12346");
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource1, resource2)).isNegative();
+    assertThat(ReconcilerUtilsInternal.compareResourceVersions(resource2, resource1)).isPositive();
+  }
+
+  private HasMetadata createResourceWithVersion(String resourceVersion) {
+    return new PodBuilder()
+        .withMetadata(
+            new ObjectMetaBuilder()
+                .withName("test-pod")
+                .withNamespace("default")
+                .withResourceVersion(resourceVersion)
+                .build())
+        .build();
   }
 }
